Hack reskillable requirement #38
This commit adds Reskillable's method of testing requirements directly to the ReskillableRequirement class. This was done to test that integration with the mod will work once they get their end sorted.

This code should not remain longer than absolutely necessary as it is hacky and is hard-coded for only two types of requirements. It will not work as expected if more requirement types are added.