Fix #108, tables not working with designer's table
I have an interface ITileEntityDesigner which is implemented by all my designer table tile entities. These TE's also extend the class TileEntity, obviously. The method boolean ITileEntityDesigner#isInvalid() is then satisfied by the super class method boolean TileEntity#isInvalid().

The method that satisfies the abstract requirement of the interface is, however, re-obfuscated at compile time because the method is part of the obfuscated MC codebase. This means that when the method is called at runtime, there is no longer an implementation of the method to back the abstract signature, and the AbstractMethodError is thrown.